solution_with_selenium/requirements.txt
django-environ==0.9.0
selenium==4.9.1